chore: Update npm dependencies to latest versions
This commit updates the npm dependencies in the `package.json` file to their latest stable versions. Specifically, the `@supabase/supabase-js` dependency is updated from version 2.39.8 to 2.44.3, and the `next` and `react-dom` dependencies are updated to version 14.2.5 and 18.3.1 respectively. Keeping the dependencies up to date ensures compatibility with the latest features and bug fixes.